feat: typeing 폴더 정리
