+++ docker-compose.yml
@@ -12,7 +12,7 @@
         - redis
       environment:
         spring.datasource.driverClass: 'com.mysql.jdbc.Driver'
-        spring.datasource.url: 'jdbc:mysql://localhost:3306/every_pet?serverTimezone=Asia/Seoul'
+        spring.datasource.url: 'jdbc:mysql://mysql:3306/every-pet?serverTimezone=Asia/Seoul'
         spring.datasource.username: 'root'
         spring.datasource.password: 'root'
         redis.host: 'redis'
