+++ frontend/src/components/common/DetailedCategory.module.css
@@ -3,31 +3,32 @@
     flex-direction: row;
     gap: 10px;
     margin-bottom: 40px;
-    overflow-x: auto; /* 가로 스크롤을 가능하게 함 */
+    overflow-x: auto;
     padding: 0 20px;
 }
 
 .item {
-    background-color: #ffffff; /* Light grey background for visibility */
-    border: 1px solid #ccc; /* Border for grid items */
-    padding: 20px; /* Padding inside grid items */
-    text-align: center; /* Center text inside grid items */
-    min-width: 120px; /* 최소 너비를 설정하여 가로 스크롤 시 보이는 크기 */
+    background-color: #ffffff;
+    border: 1px solid #ccc;
+    padding: 20px;
+    text-align: center;
+    min-width: 120px;
 
     transition: all 0.2s ease;
     cursor: pointer;
 }
 
-.item:hover {
-    background-color: #a2dcff;
+.item:hover{
+    background-color: #669AFF;
+    color: white;
 }
 
 @media screen and (min-width: 769px) {
     .container {
         display: grid;
-        grid-template-columns: repeat(5, 1fr); /* Creates 5 columns of equal width */
+        grid-template-columns: repeat(5, 1fr);
         gap: 1px;
-        overflow-x: visible; /* 데스크탑에서는 가로 스크롤 숨김 */
+        overflow-x: visible;
         padding: 0;
     }
 }
