--- frontend/src/typings/MoreInformation.ts
@@ -1,4 +1,0 @@
-export interface EatWellAndEatWellItem {
-    imageUrl: string;
-}
-
