--- frontend/src/typings/agreement.ts
@@ -1,9 +1,0 @@
-export interface AgreementJoin {
-  name: string;
-  value: string;
-  children: string;
-  checked: boolean;
-  text: string;
-  showDetails: boolean;
-  agreement: AgreementJoin[];
-}
