--- frontend/src/typings/dog_more_information.ts
@@ -1,8 +1,0 @@
-export interface dogMoreInformation {
-    name: string;
-    price: string;
-    recommended: number;
-    reviewCount: number;
-    discount: number;
-    imageUrl: string
-}
