+++ frontend/src/typings/layout.ts
@@ -6,14 +6,15 @@ export interface Ranking {
 }
 
 // 페이지 카테고리
-export interface ProductcategoryItemProps {
+export interface ProductCategoryItemProps {
     category: string;
     isActive: boolean;
     onClick: (category: string) => void;
     link: string;
 }
 
-export interface SidebarProps {
+// 메가 카테고리 온/오프
+export interface MegaMenuProps {
     isOpen: boolean;
     setOpen: () => void;
     setClose: () => void;
