+++ frontend/src/typings/product.ts
@@ -1,3 +1,6 @@
+/**
+ *  상품에 대한 정보
+ * */
 export interface Product {
     productId: string;
     productName: string;
