+++ frontend/src/typings/signup.ts
@@ -12,3 +12,14 @@ export interface Join {
   };
   agreeMarketingYn: string;
 }
+
+
+export interface AgreementJoin {
+  name: string;
+  value: string;
+  children: string;
+  checked: boolean;
+  text: string;
+  showDetails: boolean;
+  agreement: AgreementJoin[];
+}
