--- frontend/src/typings/utils.ts
@@ -1,7 +1,0 @@
-// Params 인터페이스 정의
-export interface Params {
-    orderBy: string;
-    page: number;
-    pageSize: number;
-    productCategory: string;
-}
