+++ frontend/src/utils/product/fetchProductDetails.ts
